utils: Move the Os class from core to common
The code is not specific to ORT and can be useful for other modules.

Drop the single log statement because the log helper is not available in
the common module and the (potentially fixed) user home property is
implicitly dumped as part of `ORT_DATA_DIR`.
