+++ analyzer/src/main/kotlin/managers/Bower.kt
@@ -44,8 +44,8 @@ import org.ossreviewtoolkit.model.VcsType
 import org.ossreviewtoolkit.model.config.AnalyzerConfiguration
 import org.ossreviewtoolkit.model.config.RepositoryConfiguration
 import org.ossreviewtoolkit.model.jsonMapper
+import org.ossreviewtoolkit.utils.common.Os
 import org.ossreviewtoolkit.utils.core.CommandLineTool
-import org.ossreviewtoolkit.utils.core.Os
 import org.ossreviewtoolkit.utils.core.fieldNamesOrEmpty
 import org.ossreviewtoolkit.utils.core.fieldsOrEmpty
 import org.ossreviewtoolkit.utils.core.stashDirectories
