+++ analyzer/src/main/kotlin/managers/GoMod.kt
@@ -40,8 +40,8 @@ import org.ossreviewtoolkit.model.VcsType
 import org.ossreviewtoolkit.model.config.AnalyzerConfiguration
 import org.ossreviewtoolkit.model.config.RepositoryConfiguration
 import org.ossreviewtoolkit.model.orEmpty
+import org.ossreviewtoolkit.utils.common.Os
 import org.ossreviewtoolkit.utils.core.CommandLineTool
-import org.ossreviewtoolkit.utils.core.Os
 import org.ossreviewtoolkit.utils.core.log
 import org.ossreviewtoolkit.utils.core.stashDirectories
 import org.ossreviewtoolkit.utils.core.withoutSuffix
