+++ analyzer/src/main/kotlin/managers/Gradle.kt
@@ -51,7 +51,7 @@ import org.ossreviewtoolkit.model.config.AnalyzerConfiguration
 import org.ossreviewtoolkit.model.config.RepositoryConfiguration
 import org.ossreviewtoolkit.model.createAndLogIssue
 import org.ossreviewtoolkit.model.utils.DependencyGraphBuilder
-import org.ossreviewtoolkit.utils.core.Os
+import org.ossreviewtoolkit.utils.common.Os
 import org.ossreviewtoolkit.utils.core.log
 import org.ossreviewtoolkit.utils.core.temporaryProperties
