+++ cli/src/main/kotlin/OrtMain.kt
@@ -44,12 +44,12 @@ import org.apache.logging.log4j.core.config.Configurator
 import org.ossreviewtoolkit.cli.commands.*
 import org.ossreviewtoolkit.model.config.LicenseFilenamePatterns
 import org.ossreviewtoolkit.model.config.OrtConfiguration
+import org.ossreviewtoolkit.utils.common.Os
 import org.ossreviewtoolkit.utils.core.Environment
 import org.ossreviewtoolkit.utils.core.ORT_CONFIG_DIR_ENV_NAME
 import org.ossreviewtoolkit.utils.core.ORT_CONFIG_FILENAME
 import org.ossreviewtoolkit.utils.core.ORT_DATA_DIR_ENV_NAME
 import org.ossreviewtoolkit.utils.core.ORT_NAME
-import org.ossreviewtoolkit.utils.core.Os
 import org.ossreviewtoolkit.utils.core.PERFORMANCE
 import org.ossreviewtoolkit.utils.core.expandTilde
 import org.ossreviewtoolkit.utils.core.log
